% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxplate.R
\docType{package}
\name{toxplate}
\alias{toxplate}
\alias{toxplate-package}
\title{toxplate: A package for creating plate layout mappings}
\description{
The toxplate package provides functions to creates a plate layout for your experiment based on
the top-left and the bottom-right well and a possible zigzag pattern. These layout files can
be used to match the location (aka imageNr) in the tif filename to the correct well in the plate.
The resulting plate layout data can be exported and used in CellProfiler instead of putting all
tifs in separate folders.
}

